move StructureSerializer to separate module
